+++ src/utils/mod.rs
@@ -3,8 +3,5 @@
 /// Helper to retrieve display area
 pub mod display_area;
 
-/// Construct chains of objects
-pub mod object_chain;
-
 /// Rectangle extensions
 pub mod rect_helper;